requirements.txt
# Runtime deps for live_server.py and opcua_server.py
asyncua>=1.1.0,<2.0
aiohttp>=3.9.0,<4.0
PyYAML>=6.0,<7.0
